removes dead code paths
